Move SearchStore to store package from actioncreator package
