+++ feature/session/src/main/java/io/github/droidkaigi/confsched2019/session/ui/store/SearchStore.kt
@@ -1,4 +1,4 @@
-package io.github.droidkaigi.confsched2019.session.ui.actioncreator
+package io.github.droidkaigi.confsched2019.session.ui.store
 
 import androidx.lifecycle.ViewModel
 import io.github.droidkaigi.confsched2019.action.Action
